pyproject.toml
[tool.poetry]
name = "django_auth_mafiasi"
version = "0.0.30"
description = "Django authentication library for working with Mafiasi"
license = "MIT"
authors = ["Finn-Thorben Sell <dev@finn-thorben.me>"]
maintainers = ["Mafiasi-AG <mafiasi-ag@group.mafiasi.de>"]
readme = "README.md"
repository = "https://git.mafiasi.de/mafiasi-ag/django_auth_mafiasi"
documentation = "https://docs.mafiasi.de/"

exclude = ["django_auth_mafiasi/settings.py"]

[tool.poetry.dependencies]
python = "^3.7"
django = ">=3.1"
django-auth-oidc = ">=0.6"
django-configurations = { version= ">=2.2", optional = true }

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
